requirements.txt
asgiref==3.5.2
beartype==0.10.4
beautifulsoup4==4.11.1
bs4==0.0.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
flet==0.1.50
gunicorn==20.1.0
lxml==4.9.1
numpy==1.23.3
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.5.0
python-docx==0.8.11
pytz==2022.4
repath==0.9.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
websocket-client==1.3.3
